blog/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
cachetools==5.3.3
celery==5.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.11
django-ckeditor==6.7.1
django-js-asset==2.2.0
fastapi==0.110.1
free-proxy-list==0.0.3
frozenlist==1.4.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
kombu==5.3.7
loguru==0.7.0
Markdown==3.6
multidict==6.0.5
pillow==10.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
random-user-agent==1.0.1
redis==3.5.3
requests==2.31.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
starlette==0.37.2
text-unidecode==1.3
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
win32-setctime==1.1.0
yarl==1.9.4